+++ src/smc_data_class.cpp
@@ -1,4 +1,4 @@
-#include "classes.h"
+#include "smc_classes.h"
 using namespace arma;
 
 SMCData::SMCData(
